rename parent dir to BlocksInput
