+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/BlocksContent.tsx
