+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/BlocksInput.tsx
