+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/Toolbar.tsx
