+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/plugins/withLinks.ts
