+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/plugins/withStrapiSchema.ts
