+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/tests/BlocksInput.test.tsx
