+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/tests/Toolbar.test.tsx
