+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/tests/mock-schema.ts
