+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/utils/links.ts
